+++ src/twitch/handler/LevelSettingsPopup.hpp
@@ -7,17 +7,22 @@
 
 using namespace geode::prelude;
 
+class CCMenuItemToggler;
+
 class LevelSettingsPopup : public geode::Popup<>
 {
 protected:
     TextInput* m_levelInput = nullptr;
     std::string m_value;
-    std::function<void(const std::string&)> m_callback;
+    std::function<void(const std::string&, bool)> m_callback;
+    CCMenuItemToggler* m_forceToggle = nullptr;
+    bool m_forcePlay = false;
 
     bool setup() override;
     void onOpen(CCObject* sender);
     void onSave(CCObject* sender);
+    void onToggleForce(CCObject* sender);
 
 public:
-    static LevelSettingsPopup* create(const std::string& value, std::function<void(const std::string&)> callback);
+    static LevelSettingsPopup* create(const std::string& value, bool initForce, std::function<void(const std::string&, bool)> callback);
 };